lib/pythonapp/requirements.txt
motor
parsel
httpx
pillow